Disconnected the customization tutorial.
Now that the customization docs have been updated and improved, we're not sure that this tutorial adds much. We could do with a more substantial example than what is here - which some of StephP's recommended blog posts do already supply.

We might reconnect in future if the content gets more substantial.